requirements.txt
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
djongo
gunicorn
psycopg2-binary
pymongo
django-cors-headers
djangorestframework-jwt
